기타 수정

+++ src/main/java/com/nexters/global/config/security/SecurityConstants.java
@@ -5,7 +5,7 @@ package com.nexters.global.config.security;
  */
 public class SecurityConstants {
     public static final String SECRET = "SecretKeyToGenJWTs";
-    public static final long EXPIRATION_TIME = 864_000_000; // 10 days
+    public static final Integer EXPIRATION_TIME = 864_000_000; // 10 days
     public static final String TOKEN_PREFIX = "JWT ";
     public static final String HEADER_STRING = "Authorization";
     public static final String SIGN_UP_URL = "/members/signup";
